requirements.txt
python-dotenv==1.0.1
loguru==0.7.2